Snapshot 1wx1
This Release marks the first segments of the Solar-Rental program... if you are using solar2.0 console classes, We suggest importing the messageHandler packages from this repository and replacing the one from solar2.0. You will have to change a bunch of imports on all the classes that used.

So far this repository can check first time and install a few folders and files. We are working hard to complete more code to get this program up and running.