SNAPSHOT: DNU
This snapshot fixes some of the issues caused by the last update.

Please do not use this snapshot to run the code. The code is still extremely buggy in places.

This update effects 13 files total. This snapshot is just to transfer code files to a different computer. Running this code may cause the code to crash. DNU (DO NOT USE);