Snapshot 1wx1_3 (3)x4 PART 6
Console issues in UI mode are a known issue with this release. started implementing more ui Sections. Currently the Notebook and administrative functions all support ui mode. however the notebook seems to have a slight issue where all geometry is shifted to the right. This is also a known issue.
The Maintain user controller also has support for ui mode.

Please note there still is not a setting in the settings menu that allows the swap between UI mode and console mode. However you can adjust this setting in the config.properties file just set "UI" to "Enabled" or "Disabled" and restart the program. Please note there are differences between the two modes. they are not always a perfect conversion.

Added new class that does not do anything at the moment, but it will be responsible for checking settings, and seeing if any have changed since last load. if they have it will throw a warning. This class has not yet been implemented. This will likely run on a separate thread.